requirements.txt
python-telegram-bot~=20.3      # Modern async Telegram API (requires Python 3.7+)
openai~=1.0.0                 # Correct version for current async OpenAI API (not 4.0)
aiohttp~=3.9.3                 # Async HTTP server/client
aioredis~=2.1.1                # Async Redis client
python-dotenv~=0.19.0           # .env file loading
pydantic~=2.5.0                # Data validation and settings management
cryptography~=41.0.5           # Encryption/decryption (Fernet)
tenacity~=8.2.3                # Retry logic
pyyaml~=6.0.1                  # YAML parsing for knowledge base
prometheus-client~=0.17.1      # Metrics collection